chore: Move files to src folder
