+++ src/output/interval-output.js
