+++ src/p1-reader-events.js
